requirements.txt
fastapi
uvicorn
pymupdf
sentence-transformers
qdrant-client
langchain-community
pypdf
python-dotenv
